[Refactor] * tasks' files moved to their own .py script to avoid circular import error * accuracy calculation fixed
